+++ property_core/__init__.py
@@ -34,14 +34,17 @@ from property_core.yield_service import calculate_yield
 from property_core.rightmove_location import RightmoveLocationAPI
 from property_core.rightmove_scraper import fetch_listing, fetch_listings
 from property_core.models.onthemarket import OnTheMarketListing, OnTheMarketListingDetail
-from property_core.models.zoopla import ZooplaListing
+from property_core.models.zoopla import ZooplaListing, ZooplaListingDetail
 from property_core.onthemarket_location import OnTheMarketLocationAPI
 from property_core.onthemarket_scraper import (
     fetch_listing as fetch_onthemarket_listing,
     fetch_listings as fetch_onthemarket_listings,
 )
 from property_core.zoopla_location import ZooplaLocationAPI
-from property_core.zoopla_scraper import fetch_listings as fetch_zoopla_listings
+from property_core.zoopla_scraper import (
+    fetch_listing as fetch_zoopla_listing,
+    fetch_listings as fetch_zoopla_listings,
+)
 
 __all__ = [
     # Services
@@ -70,6 +73,7 @@ __all__ = [
     "fetch_listings",
     "fetch_onthemarket_listing",
     "fetch_onthemarket_listings",
+    "fetch_zoopla_listing",
     "fetch_zoopla_listings",
     "generate_insights",
     "match_epc_address",
@@ -92,4 +96,5 @@ __all__ = [
     "StampDutyResult",
     "YieldAnalysis",
     "ZooplaListing",
+    "ZooplaListingDetail",
 ]